requirements.txt
flake8==3.7.9
yapf==0.29.0
isort==4.3.21
yacs
gdown
tb-nightly
future
scipy
scikit-learn
tqdm
ftfy
regex
setuptools==59.5.0
wilds==1.2.2
tabulate
matplotlib
open_clip_torch==2.23.0 
transformers==4.35.2